alert auto disappear implemented

+++ Admin/ajax/features_facilities.php
@@ -95,9 +95,16 @@ if (isset($_POST['rem_facility'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_facility']];
 
-    $q = "DELETE FROM `facilities` WHERE `id`=?";
-    $res = delete($q, $values, 'i');
-    echo $res;
+    $pre_q = "SELECT* FROM `facilities` WHERE `id`=?";
+    $res = select($pre_q, $values, 'i');
+    $img = mysqli_fetch_assoc($res);
 
+    if (deleteImage($img['icon'], FACILITIES_FOLDER)) {
+        $q = "DELETE FROM `facilities` WHERE `id`=?";
+        $res = delete($q, $values, 'i');
+        echo $res;
+    } else {
+        echo 0;
+    }
 }
 ?>
